LayoutTests/fast/shapes/parsing/parsing-shape-inside-expected.txt

Testing parsing of the -webkit-shape-inside property.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS getCSSText("-webkit-shape-inside", "auto") is "auto"
PASS getComputedStyleValue("-webkit-shape-inside", "auto") is "auto"
PASS getCSSText("-webkit-shape-inside", "rectangle(10px, 20px, 30px, 40px)") is "rectangle(10px, 20px, 30px, 40px)"
PASS getComputedStyleValue("-webkit-shape-inside", "rectangle(10px, 20px, 30px, 40px)") is "rectangle(10px, 20px, 30px, 40px, 0px, 0px)"
PASS getCSSText("-webkit-shape-inside", "rectangle(10px, 20px, 30px, 40px, 5px)") is "rectangle(10px, 20px, 30px, 40px, 5px)"
PASS getComputedStyleValue("-webkit-shape-inside", "rectangle(10px, 20px, 30px, 40px, 5px)") is "rectangle(10px, 20px, 30px, 40px, 5px, 5px)"
PASS getCSSText("-webkit-shape-inside", "rectangle(10px, 20px, 30px, 40px, 5px, 10px)") is "rectangle(10px, 20px, 30px, 40px, 5px, 10px)"
PASS getComputedStyleValue("-webkit-shape-inside", "rectangle(10px, 20px, 30px, 40px, 5px, 10px)") is "rectangle(10px, 20px, 30px, 40px, 5px, 10px)"
PASS getCSSText("-webkit-shape-inside", "inset-rectangle(10px, 20px, 30px, 40px)") is "inset-rectangle(10px, 20px, 30px, 40px)"
PASS getComputedStyleValue("-webkit-shape-inside", "inset-rectangle(10px, 20px, 30px, 40px)") is "inset-rectangle(10px, 20px, 30px, 40px, 0px, 0px)"
PASS getCSSText("-webkit-shape-inside", "inset-rectangle(10px, 20px, 30px, 40px, 5px)") is "inset-rectangle(10px, 20px, 30px, 40px, 5px)"
PASS getComputedStyleValue("-webkit-shape-inside", "inset-rectangle(10px, 20px, 30px, 40px, 5px)") is "inset-rectangle(10px, 20px, 30px, 40px, 5px, 5px)"
PASS getCSSText("-webkit-shape-inside", "inset-rectangle(10px, 20px, 30px, 40px, 5px, 10px)") is "inset-rectangle(10px, 20px, 30px, 40px, 5px, 10px)"
PASS getComputedStyleValue("-webkit-shape-inside", "inset-rectangle(10px, 20px, 30px, 40px, 5px, 10px)") is "inset-rectangle(10px, 20px, 30px, 40px, 5px, 10px)"
PASS getCSSText("-webkit-shape-inside", "circle(10px, 20px, 30px)") is "circle(10px, 20px, 30px)"
PASS getComputedStyleValue("-webkit-shape-inside", "circle(10px, 20px, 30px)") is "circle(10px, 20px, 30px)"
PASS getCSSText("-webkit-shape-inside", "ellipse(10px, 20px, 30px, 40px)") is "ellipse(10px, 20px, 30px, 40px)"
PASS getComputedStyleValue("-webkit-shape-inside", "ellipse(10px, 20px, 30px, 40px)") is "ellipse(10px, 20px, 30px, 40px)"
PASS getCSSText("-webkit-shape-inside", "polygon(10px 20px, 30px 40px, 40px 50px)") is "polygon(nonzero, 10px 20px, 30px 40px, 40px 50px)"
PASS getComputedStyleValue("-webkit-shape-inside", "polygon(10px 20px, 30px 40px, 40px 50px)") is "polygon(nonzero, 10px 20px, 30px 40px, 40px 50px)"
PASS getCSSText("-webkit-shape-inside", "polygon(evenodd, 10px 20px, 30px 40px, 40px 50px)") is "polygon(evenodd, 10px 20px, 30px 40px, 40px 50px)"
PASS getComputedStyleValue("-webkit-shape-inside", "polygon(evenodd, 10px 20px, 30px 40px, 40px 50px)") is "polygon(evenodd, 10px 20px, 30px 40px, 40px 50px)"
PASS getCSSText("-webkit-shape-inside", "polygon(nonzero, 10px 20px, 30px 40px, 40px 50px)") is "polygon(nonzero, 10px 20px, 30px 40px, 40px 50px)"
PASS getComputedStyleValue("-webkit-shape-inside", "polygon(nonzero, 10px 20px, 30px 40px, 40px 50px)") is "polygon(nonzero, 10px 20px, 30px 40px, 40px 50px)"
PASS getCSSText("-webkit-shape-inside", "content-box") is "content-box"
PASS getComputedStyleValue("-webkit-shape-inside", "content-box") is "content-box"
PASS getCSSText("-webkit-shape-inside", "padding-box") is "padding-box"
PASS getComputedStyleValue("-webkit-shape-inside", "padding-box") is "padding-box"
PASS getCSSText("-webkit-shape-inside", "border-box") is "border-box"
PASS getComputedStyleValue("-webkit-shape-inside", "border-box") is "border-box"
PASS getCSSText("-webkit-shape-inside", "margin-box") is "margin-box"
PASS getComputedStyleValue("-webkit-shape-inside", "margin-box") is "margin-box"
PASS getCSSText("-webkit-shape-inside", "polygon(nonzero, 0px 0px, 10px 10px, 10px 0px) content-box") is "polygon(nonzero, 0px 0px, 10px 10px, 10px 0px) content-box"
PASS getComputedStyleValue("-webkit-shape-inside", "polygon(nonzero, 0px 0px, 10px 10px, 10px 0px) content-box") is "polygon(nonzero, 0px 0px, 10px 10px, 10px 0px) content-box"
PASS getCSSText("-webkit-shape-inside", "polygon(nonzero, 0px 0px, 10px 10px, 10px 0px) padding-box") is "polygon(nonzero, 0px 0px, 10px 10px, 10px 0px) padding-box"
PASS getComputedStyleValue("-webkit-shape-inside", "polygon(nonzero, 0px 0px, 10px 10px, 10px 0px) padding-box") is "polygon(nonzero, 0px 0px, 10px 10px, 10px 0px) padding-box"
PASS getCSSText("-webkit-shape-inside", "polygon(nonzero, 0px 0px, 10px 10px, 10px 0px) border-box") is "polygon(nonzero, 0px 0px, 10px 10px, 10px 0px) border-box"
PASS getComputedStyleValue("-webkit-shape-inside", "polygon(nonzero, 0px 0px, 10px 10px, 10px 0px) border-box") is "polygon(nonzero, 0px 0px, 10px 10px, 10px 0px) border-box"
PASS getCSSText("-webkit-shape-inside", "polygon(nonzero, 0px 0px, 10px 10px, 10px 0px) margin-box") is "polygon(nonzero, 0px 0px, 10px 10px, 10px 0px) margin-box"
PASS getComputedStyleValue("-webkit-shape-inside", "polygon(nonzero, 0px 0px, 10px 10px, 10px 0px) margin-box") is "polygon(nonzero, 0px 0px, 10px 10px, 10px 0px) margin-box"
PASS getCSSText("-webkit-shape-inside", "content-box polygon(nonzero, 0px 0px, 10px 10px, 10px 0px)") is "polygon(nonzero, 0px 0px, 10px 10px, 10px 0px) content-box"
PASS getComputedStyleValue("-webkit-shape-inside", "content-box polygon(nonzero, 0px 0px, 10px 10px, 10px 0px)") is "polygon(nonzero, 0px 0px, 10px 10px, 10px 0px) content-box"
PASS getCSSText("-webkit-shape-inside", "padding-box polygon(nonzero, 0px 0px, 10px 10px, 10px 0px)") is "polygon(nonzero, 0px 0px, 10px 10px, 10px 0px) padding-box"
PASS getComputedStyleValue("-webkit-shape-inside", "padding-box polygon(nonzero, 0px 0px, 10px 10px, 10px 0px)") is "polygon(nonzero, 0px 0px, 10px 10px, 10px 0px) padding-box"
PASS getCSSText("-webkit-shape-inside", "border-box polygon(nonzero, 0px 0px, 10px 10px, 10px 0px)") is "polygon(nonzero, 0px 0px, 10px 10px, 10px 0px) border-box"
PASS getComputedStyleValue("-webkit-shape-inside", "border-box polygon(nonzero, 0px 0px, 10px 10px, 10px 0px)") is "polygon(nonzero, 0px 0px, 10px 10px, 10px 0px) border-box"
PASS getCSSText("-webkit-shape-inside", "margin-box polygon(nonzero, 0px 0px, 10px 10px, 10px 0px)") is "polygon(nonzero, 0px 0px, 10px 10px, 10px 0px) margin-box"
PASS getComputedStyleValue("-webkit-shape-inside", "margin-box polygon(nonzero, 0px 0px, 10px 10px, 10px 0px)") is "polygon(nonzero, 0px 0px, 10px 10px, 10px 0px) margin-box"
PASS getCSSText("-webkit-shape-inside", "outside-shape") is "outside-shape"
PASS getComputedStyleValue("-webkit-shape-inside", "outside-shape") is "outside-shape"
PASS removeBaseURL(getCSSText("-webkit-shape-inside", "url('image')")) is "url(image)"
PASS removeBaseURL(getComputedStyleValue("-webkit-shape-inside", "url('image')")) is "url(image)"
PASS getCSSText("-webkit-shape-inside", "calc()") is ""
PASS getComputedStyleValue("-webkit-shape-inside", "calc()") is "outside-shape"
PASS getCSSText("-webkit-shape-inside", "none") is ""
PASS getComputedStyleValue("-webkit-shape-inside", "none") is "outside-shape"
PASS getCSSText("-webkit-shape-inside", "rectangle()") is ""
PASS getComputedStyleValue("-webkit-shape-inside", "rectangle()") is "outside-shape"
PASS getCSSText("-webkit-shape-inside", "rectangle(10px)") is ""
PASS getComputedStyleValue("-webkit-shape-inside", "rectangle(10px)") is "outside-shape"
PASS getCSSText("-webkit-shape-inside", "rectangle(10px, 10px)") is ""
PASS getComputedStyleValue("-webkit-shape-inside", "rectangle(10px, 10px)") is "outside-shape"
PASS getCSSText("-webkit-shape-inside", "rectangle(10px, 20px, 30px)") is ""
PASS getComputedStyleValue("-webkit-shape-inside", "rectangle(10px, 20px, 30px)") is "outside-shape"
PASS getCSSText("-webkit-shape-inside", "rectangle(10px 20px 30px 40px)") is ""
PASS getComputedStyleValue("-webkit-shape-inside", "rectangle(10px 20px 30px 40px)") is "outside-shape"
PASS getCSSText("-webkit-shape-inside", "rectangle(10px, 20px, 30px, 40px, 50px, 60px, 70px)") is ""
PASS getComputedStyleValue("-webkit-shape-inside", "rectangle(10px, 20px, 30px, 40px, 50px, 60px, 70px)") is "outside-shape"
PASS getCSSText("-webkit-shape-inside", "inset-rectangle()") is ""
PASS getComputedStyleValue("-webkit-shape-inside", "inset-rectangle()") is "outside-shape"
PASS getCSSText("-webkit-shape-inside", "inset-rectangle(10px)") is ""
PASS getComputedStyleValue("-webkit-shape-inside", "inset-rectangle(10px)") is "outside-shape"
PASS getCSSText("-webkit-shape-inside", "inset-rectangle(10px, 10px)") is ""
PASS getComputedStyleValue("-webkit-shape-inside", "inset-rectangle(10px, 10px)") is "outside-shape"
PASS getCSSText("-webkit-shape-inside", "inset-rectangle(10px, 20px, 30px)") is ""
PASS getComputedStyleValue("-webkit-shape-inside", "inset-rectangle(10px, 20px, 30px)") is "outside-shape"
PASS getCSSText("-webkit-shape-inside", "inset-rectangle(10px 20px 30px 40px)") is ""
PASS getComputedStyleValue("-webkit-shape-inside", "inset-rectangle(10px 20px 30px 40px)") is "outside-shape"
PASS getCSSText("-webkit-shape-inside", "inset-rectangle(10px, 20px, 30px, 40px, 50px, 60px, 70px)") is ""
PASS getComputedStyleValue("-webkit-shape-inside", "inset-rectangle(10px, 20px, 30px, 40px, 50px, 60px, 70px)") is "outside-shape"
PASS getCSSText("-webkit-shape-inside", "circle()") is ""
PASS getComputedStyleValue("-webkit-shape-inside", "circle()") is "outside-shape"
PASS getCSSText("-webkit-shape-inside", "circle(10px)") is ""
PASS getComputedStyleValue("-webkit-shape-inside", "circle(10px)") is "outside-shape"
PASS getCSSText("-webkit-shape-inside", "circle(10px, 20px)") is ""
PASS getComputedStyleValue("-webkit-shape-inside", "circle(10px, 20px)") is "outside-shape"
PASS getCSSText("-webkit-shape-inside", "circle(10px 20px 30px)") is ""
PASS getComputedStyleValue("-webkit-shape-inside", "circle(10px 20px 30px)") is "outside-shape"
PASS getCSSText("-webkit-shape-inside", "circle(10px, 20px, 30px, 40px)") is ""
PASS getComputedStyleValue("-webkit-shape-inside", "circle(10px, 20px, 30px, 40px)") is "outside-shape"
PASS getCSSText("-webkit-shape-inside", "ellipse()") is ""
PASS getComputedStyleValue("-webkit-shape-inside", "ellipse()") is "outside-shape"
PASS getCSSText("-webkit-shape-inside", "ellipse(10px)") is ""
PASS getComputedStyleValue("-webkit-shape-inside", "ellipse(10px)") is "outside-shape"
PASS getCSSText("-webkit-shape-inside", "ellipse(10px, 20px)") is ""
PASS getComputedStyleValue("-webkit-shape-inside", "ellipse(10px, 20px)") is "outside-shape"
PASS getCSSText("-webkit-shape-inside", "ellipse(10px, 20px, 30px)") is ""
PASS getComputedStyleValue("-webkit-shape-inside", "ellipse(10px, 20px, 30px)") is "outside-shape"
PASS getCSSText("-webkit-shape-inside", "ellipse(10px 20px 30px 40px)") is ""
PASS getComputedStyleValue("-webkit-shape-inside", "ellipse(10px 20px 30px 40px)") is "outside-shape"
PASS getCSSText("-webkit-shape-inside", "polygon()") is ""
PASS getComputedStyleValue("-webkit-shape-inside", "polygon()") is "outside-shape"
PASS getCSSText("-webkit-shape-inside", "polygon(evenodd 10px 20px, 30px 40px, 40px 50px)") is ""
PASS getComputedStyleValue("-webkit-shape-inside", "polygon(evenodd 10px 20px, 30px 40px, 40px 50px)") is "outside-shape"
PASS getCSSText("-webkit-shape-inside", "polygon(nonzero 10px 20px, 30px 40px, 40px 50px)") is ""
PASS getComputedStyleValue("-webkit-shape-inside", "polygon(nonzero 10px 20px, 30px 40px, 40px 50px)") is "outside-shape"
PASS getCSSText("-webkit-shape-inside", "polygon(nonzero)") is ""
PASS getComputedStyleValue("-webkit-shape-inside", "polygon(nonzero)") is "outside-shape"
PASS getCSSText("-webkit-shape-inside", "polygon(evenodd)") is ""
PASS getComputedStyleValue("-webkit-shape-inside", "polygon(evenodd)") is "outside-shape"
PASS getCSSText("-webkit-shape-inside", "polygon(10px)") is ""
PASS getComputedStyleValue("-webkit-shape-inside", "polygon(10px)") is "outside-shape"
PASS getCSSText("-webkit-shape-inside", "polygon(nonzero,10px)") is ""
PASS getComputedStyleValue("-webkit-shape-inside", "polygon(nonzero,10px)") is "outside-shape"
PASS getCSSText("-webkit-shape-inside", "polygon(evenodd,12px)") is ""
PASS getComputedStyleValue("-webkit-shape-inside", "polygon(evenodd,12px)") is "outside-shape"
PASS getCSSText("-webkit-shape-inside", "polygon(10px, 20px, 30px, 40px, 40px, 50px)") is ""
PASS getComputedStyleValue("-webkit-shape-inside", "polygon(10px, 20px, 30px, 40px, 40px, 50px)") is "outside-shape"
PASS getCSSText("-webkit-shape-inside", "content-box content-box") is ""
PASS getComputedStyleValue("-webkit-shape-inside", "content-box content-box") is "outside-shape"
PASS getCSSText("-webkit-shape-inside", "polygon(nonzero, 0 0, 10px 10px, 10px 0) polygon(nonzero, 0 0, 10px 10px, 10px 0)") is ""
PASS getComputedStyleValue("-webkit-shape-inside", "polygon(nonzero, 0 0, 10px 10px, 10px 0) polygon(nonzero, 0 0, 10px 10px, 10px 0)") is "outside-shape"
PASS getCSSText("-webkit-shape-inside", "content-box polygon(nonzero, 0 0, 10px 10px, 10px 0) content-box") is ""
PASS getComputedStyleValue("-webkit-shape-inside", "content-box polygon(nonzero, 0 0, 10px 10px, 10px 0) content-box") is "outside-shape"
PASS getCSSText("-webkit-shape-inside", "auto content-box") is ""
PASS getComputedStyleValue("-webkit-shape-inside", "auto content-box") is "outside-shape"
PASS getCSSText("-webkit-shape-inside", "url('shape.svg') content-box") is ""
PASS getComputedStyleValue("-webkit-shape-inside", "url('shape.svg') content-box") is "outside-shape"
PASS getCSSText("-webkit-shape-inside", "url('shape.svg') polygon(nonzero, 0 0, 10px 10px, 10px 0)") is ""
PASS getComputedStyleValue("-webkit-shape-inside", "url('shape.svg') polygon(nonzero, 0 0, 10px 10px, 10px 0)") is "outside-shape"
PASS getParentAndChildComputedStylesString("-webkit-shape-inside", "auto", "rectangle(10px, 20px, 30px, 40px)") is "parent: auto, child: rectangle(10px, 20px, 30px, 40px, 0px, 0px)"
PASS getParentAndChildComputedStylesString("-webkit-shape-inside", "outside-shape", "rectangle(10px, 20px, 30px, 40px)") is "parent: outside-shape, child: rectangle(10px, 20px, 30px, 40px, 0px, 0px)"
PASS getParentAndChildComputedStylesString("-webkit-shape-inside", "rectangle(10px, 20px, 30px, 40px)", "initial") is "parent: rectangle(10px, 20px, 30px, 40px, 0px, 0px), child: outside-shape"
PASS getParentAndChildComputedStylesString("-webkit-shape-inside", "rectangle(10px, 20px, 30px, 40px)", "") is "parent: rectangle(10px, 20px, 30px, 40px, 0px, 0px), child: outside-shape"
PASS getParentAndChildComputedStylesString("-webkit-shape-inside", "rectangle(10px, 20px, 30px, 40px)", "inherit") is "parent: rectangle(10px, 20px, 30px, 40px, 0px, 0px), child: rectangle(10px, 20px, 30px, 40px, 0px, 0px)"
PASS getParentAndChildComputedStylesString("-webkit-shape-inside", "", "inherit") is "parent: outside-shape, child: outside-shape"
PASS getParentAndChildComputedStylesString("-webkit-shape-inside", "auto", "inherit") is "parent: auto, child: auto"
PASS successfullyParsed is true

TEST COMPLETE
